requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.3
bitarray==3.5.0
blinker==1.9.0
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
click-default-group==1.2.4
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.5
cycler==0.12.1
cytoolz==1.0.1
decorator==5.2.1
distro==1.9.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
executing==2.2.0
fastapi==0.116.1
ffmpeg-python==0.2.0
firebase-admin==6.9.0
Flask==3.1.1
fonttools==4.56.0
frozenlist==1.7.0
ftfy==6.3.1
future==1.0.0
gallery_dl==1.29.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hexbytes==1.3.1
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
humanize==4.12.1
hyperframe==5.2.0
idna==2.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
library==3.0.69
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
mutagen==1.47.0
narwhals==1.46.0
natsort==8.4.0
npm==0.1.1
numpy==2.2.4
openai==1.95.1
optional-django==0.1.0
packaging==24.2
pandas==2.3.0
parsimonious==0.10.0
parso==0.8.4
pillow==11.1.0
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pure_eval==0.2.3
puremagic==1.28
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.22.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyreadline3==3.5.4
pysubs2==1.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-mpv-jsonipc==1.2.0
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
pywin32==310
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rlp==4.1.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
screeninfo==0.8.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.3
sqlite-fts4==1.0.3
sqlite-utils==3.38
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.47.1
streamlit==1.46.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinytag==2.1.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
web3==7.12.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
yt-dlp==2025.2.19
